update subsidio_sppo_apuracao flow
