+++ pipelines/rj_smtr/projeto_subsidio_sppo/tasks.py
@@ -8,31 +8,3 @@ import pandas as pd
 from prefect import task
 
 from pipelines.utils.tasks import log, get_now_date
-
-
-@task
-def get_run_dates(date_range_start: str, date_range_end: str) -> List:
-    """
-    Generates a list of dates between date_range_start and date_range_end.
-    """
-    if (date_range_start is False) or (date_range_end is False):
-        dates = [{"run_date": get_now_date.run()}]
-    else:
-        dates = [
-            {"run_date": d.strftime("%Y-%m-%d")}
-            for d in pd.date_range(start=date_range_start, end=date_range_end)
-        ]
-    log(f"Will run the following dates: {dates}")
-    return dates
-
-
-@task
-def get_join_dict(dict_list: list, new_dict: dict) -> List:
-    """
-    Updates a list of dictionaries with a new dictionary.
-    """
-    for dict_temp in dict_list:
-        dict_temp.update(new_dict)
-
-    log(f"get_join_dict: {dict_list}")
-    return dict_list
